Combine Routing and Serialization
